pre-review cleanup
